Move NbtUtil to Screen API module
