+++ ScreenNBT/src/main/java/opekope2/optigui/screen_nbt/mixin/LoomMenuMixin.java
@@ -9,7 +9,7 @@ import net.minecraft.world.Container;
 import net.minecraft.world.inventory.LoomMenu;
 import net.minecraft.world.level.block.entity.BannerPattern;
 import opekope2.optigui.screen_api.util.INbtConvertible;
-import opekope2.optigui.screen_nbt.util.NbtUtil;
+import opekope2.optigui.screen_api.util.NbtUtil;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
