+++ ScreenNBT/src/main/java/opekope2/optigui/screen_nbt/mixin/MerchantMenuMixin.java
@@ -6,7 +6,7 @@ import net.minecraft.world.inventory.MerchantContainer;
 import net.minecraft.world.inventory.MerchantMenu;
 import net.minecraft.world.item.trading.MerchantOffers;
 import opekope2.optigui.screen_api.util.INbtConvertible;
-import opekope2.optigui.screen_nbt.util.NbtUtil;
+import opekope2.optigui.screen_api.util.NbtUtil;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
