--- ScreenNBT/src/main/java/opekope2/optigui/screen_nbt/util/package-info.java
@@ -1,4 +1,0 @@
-@NullMarked
-package opekope2.optigui.screen_nbt.util;
-
-import org.jspecify.annotations.NullMarked;
